backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
catboost
requests
python-dotenv
openai
